Login and signup buttons added
