+++ src/common/buttons/underline-button.module.scss
@@ -1,0 +1,7 @@
+.customUnderlineButton {
+  height: 3em;
+  border-bottom: solid;
+  border-width: 3px;
+  border-color: #79c7ee;
+  font-weight: bold;
+}
